+++ cmd/strike/container.go
@@ -14,7 +14,7 @@ import (
 	"time"
 
 	"github.com/jonathanung/strike-cli/internal/config"
-	"github.com/jonathanung/strike-cli/internal/container"
+	"github.com/jonathanung/strike-cli/internal/integrate/container"
 	"github.com/jonathanung/strike-cli/internal/version"
 )